# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "[BDH] Implement for support project Billing Upgrade.",
        "title": "[BDH] Swagger APIs",
        "contact": {},
        "version": "00.01"
    },
    "basePath": "/",
    "paths": {
        "/bdhgateway/getcreditlimitwithperformanceindicators": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bdh-gateway-controller"
                ],
                "summary": "bdhgateway",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-Channel",
                        "name": "X-Channel",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "X-GatewayType",
                        "name": "X-GatewayType",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "X-LegacyUsername",
                        "name": "X-LegacyUsername",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "X-Username",
                        "name": "X-Username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "requestInfo",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apimodel.GetCreditLimitWithPerformanceIndicatorsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/apimodel.GetCreditLimitWithPerformanceIndicatorsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apimodel.HttpErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apimodel.HttpErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apimodel.HttpErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/apimodel.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/bdhgateway/getusage": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bdh-gateway-controller"
                ],
                "summary": "bdhgateway",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-Channel",
                        "name": "X-Channel",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "X-GatewayType",
                        "name": "X-GatewayType",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "X-LegacyUsername",
                        "name": "X-LegacyUsername",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "X-Username",
                        "name": "X-Username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "requestInfo",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apimodel.GetUsageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/apimodel.GetUsageResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apimodel.HttpErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apimodel.HttpErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apimodel.HttpErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/apimodel.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/bdhgateway/usageconsumption": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bdh-gateway-controller"
                ],
                "summary": "bdhgateway",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-Channel",
                        "name": "X-Channel",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "X-GatewayType",
                        "name": "X-GatewayType",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "X-LegacyUsername",
                        "name": "X-LegacyUsername",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "X-Username",
                        "name": "X-Username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "requestInfo",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apimodel.UsageConsumptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/apimodel.UsageConsumptionResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apimodel.HttpErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apimodel.HttpErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apimodel.HttpErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/apimodel.HttpErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "apimodel.AttributeList": {
            "type": "object",
            "properties": {
                "attributeName": {
                    "type": "string"
                },
                "attributeValue": {
                    "type": "string"
                }
            }
        },
        "apimodel.BackendResponseInfoArray": {
            "type": "object",
            "properties": {
                "apiName": {
                    "type": "string"
                },
                "errorCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "system": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "apimodel.BackendResponseList": {
            "type": "object",
            "properties": {
                "backendResponseInfoArray": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apimodel.BackendResponseInfoArray"
                    }
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "apimodel.CustomerCyclesInfos": {
            "type": "object",
            "properties": {
                "customerNo": {
                    "type": "integer"
                },
                "customerPartitionId": {
                    "type": "integer"
                },
                "cycleCode": {
                    "type": "integer"
                },
                "cycleInstance": {
                    "type": "integer"
                },
                "cycleSeqNo": {
                    "type": "integer"
                },
                "cycleYear": {
                    "type": "integer"
                },
                "endDate": {
                    "type": "string"
                },
                "runDate": {
                    "type": "object"
                },
                "startDate": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "apimodel.CycleInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "month": {
                    "type": "string"
                },
                "year": {
                    "type": "string"
                }
            }
        },
        "apimodel.Data": {
            "type": "object",
            "properties": {
                "accountInfo": {
                    "$ref": "#/definitions/apimodel.AccountInfo"
                },
                "customerCyclesInfos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apimodel.CustomerCyclesInfos"
                    }
                },
                "nameAddressInfoList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apimodel.NameAddressInfoList"
                    }
                },
                "performanceIndicatorInfos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apimodel.PerformanceIndicatorInfos"
                    }
                }
            }
        },
        "apimodel.ErrorMessageInfo": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string"
                },
                "errorMessage": {
                    "type": "string"
                }
            }
        },
        "apimodel.FlexibleCycleInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "month": {
                    "type": "string"
                },
                "year": {
                    "type": "string"
                }
            }
        },
        "apimodel.GetCreditLimitWithPerformanceIndicatorsList": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apimodel.Results"
                    }
                }
            }
        },
        "apimodel.GetCreditLimitWithPerformanceIndicatorsRequest": {
            "type": "object",
            "required": [
                "requestList"
            ],
            "properties": {
                "requestList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apimodel.RequestList"
                    }
                }
            }
        },
        "apimodel.GetCreditLimitWithPerformanceIndicatorsResponse": {
            "type": "object",
            "properties": {
                "backendResponseList": {
                    "$ref": "#/definitions/apimodel.BackendResponseList"
                },
                "errorCode": {
                    "type": "string"
                },
                "getCreditLimitWithPerformanceIndicatorsList": {
                    "$ref": "#/definitions/apimodel.GetCreditLimitWithPerformanceIndicatorsList"
                },
                "message": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "apimodel.GetRatedEventsListInfoResponse": {
            "type": "object",
            "properties": {
                "ratedEventsInfoMList": {
                    "$ref": "#/definitions/apimodel.RatedEventsInfoList"
                }
            }
        },
        "apimodel.GetRatedPerformanceIndicatorsList": {
            "type": "object",
            "properties": {
                "attributeList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apimodel.AttributeList"
                    }
                },
                "pIType": {
                    "type": "string"
                },
                "performanceIndicatorID": {
                    "$ref": "#/definitions/apimodel.PerformanceIndicatorID"
                }
            }
        },
        "apimodel.GetUsageRequest": {
            "type": "object",
            "required": [
                "ban",
                "correlatedId",
                "customerId",
                "cycleCode",
                "cycleInstance",
                "cycleYear"
            ],
            "properties": {
                "SubscriberId": {
                    "type": "string"
                },
                "ban": {
                    "type": "string"
                },
                "callerUuid": {
                    "type": "string"
                },
                "correlatedId": {
                    "type": "string"
                },
                "customerId": {
                    "type": "string"
                },
                "cycleCode": {
                    "type": "string"
                },
                "cycleInstance": {
                    "type": "string"
                },
                "cycleYear": {
                    "type": "string"
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                }
            }
        },
        "apimodel.GetUsageResponse": {
            "type": "object",
            "properties": {
                "backendResponseList": {
                    "$ref": "#/definitions/apimodel.BackendResponseList"
                },
                "errorCode": {
                    "type": "string"
                },
                "getRatedEventsListInfoResponse": {
                    "$ref": "#/definitions/apimodel.GetRatedEventsListInfoResponse"
                },
                "message": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "apimodel.HttpErrorResponse": {
            "type": "object",
            "properties": {
                "@timestamp": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "apimodel.NameAddressInfoList": {
            "type": "object",
            "properties": {
                "addressInfo": {
                    "$ref": "#/definitions/apimodel.AddressInfo"
                },
                "nameInfo": {
                    "$ref": "#/definitions/apimodel.NameInfo"
                }
            }
        },
...
        "apimodel.PerformanceIndicatorID": {
            "type": "object",
            "properties": {
                "agreementId": {
                    "type": "string"
                },
                "cycleInfo": {
                    "$ref": "#/definitions/apimodel.CycleInfo"
                },
                "flexibleCycleInfo": {
                    "$ref": "#/definitions/apimodel.FlexibleCycleInfo"
                },
                "offerInstance": {
                    "type": "string"
                },
                "partitionId": {
                    "type": "string"
                },
                "performanceIndicatorId": {
                    "type": "string"
                }
            }
        },
...
        "apimodel.RatedEventsInfoList": {
            "type": "object",
            "properties": {
                "errorMessage": {
                    "$ref": "#/definitions/apimodel.ErrorMessageInfo"
                },
                "ratedEventsInfoM": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apimodel.RatedEventsInfo"
                    }
                },
                "totalSize": {
                    "type": "string"
                },
                "transactionId": {
                    "type": "string"
                }
            }
        },
        "apimodel.RequestList": {
            "type": "object",
            "properties": {
                "accountNo": {
                    "type": "integer"
                },
                "subscriberNo": {
                    "type": "integer"
                }
            }
        },
        "apimodel.Results": {
            "type": "object",
            "properties": {
                "accountNo": {
                    "type": "integer"
                },
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/apimodel.Data"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subscriberNo": {
                    "type": "integer"
                }
            }
        },
        "apimodel.UsageConsumptionRequest": {
            "type": "object",
            "required": [
                "agreementId",
                "correlatedId",
                "customerId",
                "cycleCode",
                "cycleMonth",
                "cycleYear",
                "externalStructure"
            ],
            "properties": {
                "agreementId": {
                    "type": "string"
                },
                "callerUuid": {
                    "type": "string"
                },
                "correlatedId": {
                    "type": "string"
                },
                "customerId": {
                    "type": "string"
                },
                "cycleCode": {
                    "type": "string"
                },
                "cycleMonth": {
                    "type": "string"
                },
                "cycleYear": {
                    "type": "string"
                },
                "externalStructure": {
                    "type": "string"
                },
                "pageNumber": {
                    "type": "string"
                },
                "pageSize": {
                    "type": "string"
                }
            }
        },
        "apimodel.UsageConsumptionResponse": {
            "type": "object",
            "properties": {
                "backendResponseList": {
                    "$ref": "#/definitions/apimodel.BackendResponseList"
                },
                "errorCode": {
                    "type": "string"
                },
                "getRatedPerformanceIndicatorsList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apimodel.GetRatedPerformanceIndicatorsList"
                    }
                },
                "message": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}